move resolver to separate module
